+++ tests/bfabric_app_runner/specs/test_outputs_spec.py
@@ -35,7 +35,8 @@ def serialized() -> str:
   store_folder_path: null
   type: bfabric_copy_resource
   update_existing: 'no'
-- has_header: true
+- format: csv
+  has_header: true
   invalid_characters: ''
   local_path: local_path
   name: null